Fix type conversions

+++ src/shared/dictionary.c
@@ -141,7 +141,7 @@ unsigned dictionary_hash(const char * key)
   dictionary, give size=0.
  */
 /*--------------------------------------------------------------------------*/
-dictionary * dictionary_new(int size)
+dictionary * dictionary_new(size_t size)
 {
     dictionary  *   d ;
